+++ setup/dashboard/db/branding-m.php
@@ -1,4 +1,6 @@
 <?php
+// SPDX-License-Identifier: GPL-3.0-or-later
+
 assert(isset($branch));
 ?>
 <ul class="dropdown-menu pull-right">
@@ -10,6 +12,6 @@ assert(isset($branch));
   <?php } ?>
   </li>
   <li style="border-top: 1px solid #444">
-    <a href="https://github.com/amefs/quickbox-arm64/issues/new" target="_blank" rel="noopener"><i class="fa fa-warning text-warning"></i><?php echo T('ISSUE_REPORT_TXT'); ?></a>
+    <a href="https://github.com/amefs/quickbox-arm64/issues/new" target="_blank" rel="noopenner noreferrer"><i class="fa fa-warning text-warning"></i><?php echo T('ISSUE_REPORT_TXT'); ?></a>
   </li>
 </ul>