+++ setup/dashboard/widgets/service_control.php
@@ -1,21 +1,23 @@
 <?php
-  require_once($_SERVER['DOCUMENT_ROOT'].'/inc/localize.php');
-  require_once($_SERVER['DOCUMENT_ROOT'].'/inc/info.package.php');
-  assert(isset($packageList));
+// SPDX-License-Identifier: GPL-3.0-or-later
 
-  /**
-   * @param string $service
-   * @param string $username
-   *
-   * @return string
-   */
-  function isServiceEnabled($service, $username) {
-      if (file_exists('/etc/systemd/system/multi-user.target.wants/'.$service.'@'.$username.'.service') || file_exists('/etc/systemd/system/multi-user.target.wants/'.$service.'.service')) {
-          return ' <div class="toggle-wrapper text-center"><div onclick="serviceUpdateHandler(event)" class="toggle-en toggle-light primary" data-service="'.$service.'" data-operation="stop,disable"></div></div>';
-      } else {
-          return ' <div class="toggle-wrapper text-center"><div onclick="serviceUpdateHandler(event)" class="toggle-dis toggle-light primary" data-service="'.$service.'" data-operation="enable,restart"></div></div>';
-      }
-  }
+require_once($_SERVER['DOCUMENT_ROOT'].'/inc/localize.php');
+require_once($_SERVER['DOCUMENT_ROOT'].'/inc/info.package.php');
+assert(isset($packageList));
+
+/**
+ * @param string $service
+ * @param string $username
+ *
+ * @return string
+ */
+function isServiceEnabled($service, $username) {
+    if (file_exists('/etc/systemd/system/multi-user.target.wants/'.$service.'@'.$username.'.service') || file_exists('/etc/systemd/system/multi-user.target.wants/'.$service.'.service')) {
+        return ' <div class="toggle-wrapper text-center"><div onclick="serviceUpdateHandler(event)" class="toggle-en toggle-light primary" data-service="'.$service.'" data-operation="stop,disable"></div></div>';
+    } else {
+        return ' <div class="toggle-wrapper text-center"><div onclick="serviceUpdateHandler(event)" class="toggle-dis toggle-light primary" data-service="'.$service.'" data-operation="enable,restart"></div></div>';
+    }
+}
 ?>
 
 <!--SERVICE CONTROL CENTER-->
@@ -53,7 +55,7 @@
         <?php
             }
         }
-        ?>
+?>
         </tbody>
       </table>
     </div><!-- table-responsive -->